Don't report partial results when aborted
Summary:
This has sometimes been a bug for a while, but I think it got worse when
we moved to using `testresults_util` to report partial results, since
`notify` handles things a bit better.  Now, we just skip result-analysis
entirely if we were aborted.  (I suspect this may still be imperfect
because sometimes the build status is confusing which is why we had bugs
before the change, but hopefully it will be good enough.)

Test Plan:
- ran tests, using replay to put an `echo(currentBuild.result)` in near
  here, checked that it was indeed 'ABORTED' on abort.
- run tests, abort after some workers finish, don't get test results.

Reviewers: csilvers, jacqueline

Reviewed By: jacqueline

Subscribers: Kai, mita

Differential Revision: https://phabricator.khanacademy.org/D49491

+++ jobs/webapp-test.groovy
@@ -342,6 +342,13 @@ def determineSplitsAndRunTests() {
 
 def analyzeResults() {
    withTimeout('5m') {
+      if (currentBuild.result == 'ABORTED') {
+         // No need to report the results in the case of abort!  They will
+         // likely be more confusing than useful.
+         echo('We were aborted; no need to report results.');
+         return;
+      }
+
       sh("rm -f test-results.*.pickle");
       for (def i = 0; i < NUM_WORKER_MACHINES; i++) {
          try {
